// inventory-web/src/swagger-client/model/registerRequest.ts
/**
 * Elton's Inventory
 * Reboot
 *
 * OpenAPI spec version: 1.0.11
 * Contact: nhlana.2@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RegisterRequest { 
    email: string;
    password: string;
    name: string;
}